requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-decouple==3.8
Pillow==9.5.0
django-filter==23.1